Refactor: Split attio-operations.ts into focused modules (Issue #88)
- Created modular structure for API operations
- Separated search, CRUD, notes, lists, batch, and retry logic
- Maintained backward compatibility through index.ts
- Updated all imports throughout codebase
- Removed original 1,169-line attio-operations.ts file

Closes #88
