+++ src/api/operations/index.ts
@@ -1,0 +1,65 @@
+/**
+ * Central export point for all API operations
+ * Maintains backward compatibility with the original attio-operations.ts
+ */
+
+// Re-export all retry functionality
+export {
+  callWithRetry,
+  calculateRetryDelay,
+  isRetryableError,
+  DEFAULT_RETRY_CONFIG
+} from './retry.js';
+export type { RetryConfig } from './retry.js';
+
+// Re-export all types
+export type {
+  BatchRequestItem,
+  BatchItemResult,
+  BatchResponse,
+  BatchConfig,
+  ListEntryFilter,
+  ListEntryFilters
+} from './types.js';
+
+// Re-export all search operations
+export {
+  searchObject,
+  advancedSearchObject,
+  listObjects
+} from './search.js';
+
+// Re-export all CRUD operations
+export {
+  getObjectDetails,
+  createRecord,
+  getRecord,
+  updateRecord,
+  deleteRecord,
+  listRecords
+} from './crud.js';
+
+// Re-export all notes operations
+export {
+  getObjectNotes,
+  createObjectNote
+} from './notes.js';
+
+// Re-export all list operations
+export {
+  getAllLists,
+  getListDetails,
+  getListEntries,
+  addRecordToList,
+  removeRecordFromList
+} from './lists.js';
+
+// Re-export all batch operations
+export {
+  batchCreateRecords,
+  batchUpdateRecords,
+  executeBatchOperations,
+  batchSearchObjects,
+  batchGetObjectDetails,
+  DEFAULT_BATCH_CONFIG
+} from './batch.js';
