+++ src/objects/records.ts
@@ -12,7 +12,7 @@ import {
   batchUpdateRecords,
   BatchConfig,
   BatchResponse
-} from "../api/attio-operations.js";
+} from "../api/operations/index.js";
 import {
   ResourceType,
   AttioRecord,
